requirements.txt
djangorestframework~=3.12.2
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-import-export==2.5.0
django-restframework-apiview==1.4.2
djangorestframework==3.12.2
et-xmlfile==1.0.1
google-play-scraper==0.1.2
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
psycopg2==2.8.6
pytz==2020.5
PyYAML==5.4
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0

